feat: useNotification 훅 작성 및 적용
